requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.4
openpyxl==3.1.2

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10

# Vector embeddings and search
langchain-nvidia-ai-endpoints==0.0.11
faiss-cpu==1.7.4
numpy==1.24.4

# LLM integrations
openai==1.3.7
google-generativeai==0.3.2

# LangGraph agent
langgraph==0.0.62
langchain-core==0.1.0
langchain-openai==0.0.2

# Web scraping and parsing
requests==2.31.0
beautifulsoup4==4.12.2

# Environment and utilities
python-dotenv==1.0.0
pytz==2023.3

# Logging and monitoring
websockets==12.0